Remade the fraction constructors to be much more correct in their operations. Fraction values now always conserve signs. Check fraction constructor unit tests

+++ Types_of_Numbers/utility_functions.h
@@ -3,15 +3,15 @@
 class utility_functions
 {
 public:
-	//adds characters not found from the first list to the second list
+	//adds any type values not found from the first list to the second list
 	template <class MyType>
 	std::vector<MyType> static add_missing(std::vector<MyType> const& list1, std::vector<MyType> const& list2)
 	{
 		std::vector<MyType> temp = list2;
 		bool missing = true;
-		for (int i = 0; i < list1.size(); i++, missing = true)
+		for (size_t i = 0; i < list1.size(); i++, missing = true)
 		{
-			for (int j = 0; j < temp.size(); j++)
+			for (size_t j = 0; j < temp.size(); j++)
 				if (list1[i] == temp[j])
 					missing = false;
 			if (missing)
